wip; now some data are sent; but not persisted i guess

+++ services/recom-backend/src/main/java/com/recom/dto/map/scanner/MapTopographyChunkScanRequestDto.java
@@ -25,11 +25,11 @@ public class MapTopographyChunkScanRequestDto {
     @NotEmpty
     @Schema
     @JsonProperty()
-    private Integer chunkCoordinateX;
+    private Long chunkCoordinateX;
 
     @NotEmpty
     @Schema
     @JsonProperty()
-    private Integer chunkCoordinateY;
+    private Long chunkCoordinateY;
 
 }
